+++ docs/ReleaseNotes.md
@@ -2,7 +2,8 @@
 
 ## [1.6.0](https://www.nuget.org/packages/WeihanLi.Npoi/1.6.0)
 
-- add hidden property support
+- add shadow property support
+- add version info when export `*.xlsx` excel
 
 ## [1.5.0](https://www.nuget.org/packages/WeihanLi.Npoi/1.5.0)